Make some lines shorter.
This makes it easier to read on smaller screens (one of which I am
currently using right now, hence my notice of it). It's also generally
considered to be good practice not to go over 80 characters, so I
wrapped most of the comments down to 80-character lines with
4-space tab-size. I also wrapped some comments down so that they will
fit more nicely over the lines they concern. To me, code looks more
beautiful if it's more rectangular than if you have a long comment
corresponding to a short line; splitting it up over multiple lines of
more equal length is likely to result in prettier stuff.